+++ pyflakes/test/test_type_annotations.py
@@ -17,22 +17,22 @@ class TestTypeAnnotations(TestCase):
         from typing import overload
 
         @overload
-        def f(s):  # type: (None) -> None
+        def f(s: None) -> None:
             pass
 
         @overload
-        def f(s):  # type: (int) -> int
+        def f(s: int) -> int:
             pass
 
         def f(s):
             return s
 
         @typing.overload
-        def g(s):  # type: (None) -> None
+        def g(s: None) -> None:
             pass
 
         @typing.overload
-        def g(s):  # type: (int) -> int
+        def g(s: int) -> int:
             pass
 
         def g(s):
@@ -46,22 +46,22 @@ class TestTypeAnnotations(TestCase):
         from typing_extensions import overload
 
         @overload
-        def f(s):  # type: (None) -> None
+        def f(s: None) -> None:
             pass
 
         @overload
-        def f(s):  # type: (int) -> int
+        def f(s: int) -> int:
             pass
 
         def f(s):
             return s
 
         @typing_extensions.overload
-        def g(s):  # type: (None) -> None
+        def g(s: None) -> None:
             pass
 
         @typing_extensions.overload
-        def g(s):  # type: (int) -> int
+        def g(s: int) -> int:
             pass
 
         def g(s):
@@ -74,11 +74,11 @@ class TestTypeAnnotations(TestCase):
         from typing import overload
 
         @overload
-        async def f(s):  # type: (None) -> None
+        async def f(s: None) -> None:
             pass
 
         @overload
-        async def f(s):  # type: (int) -> int
+        async def f(s: int) -> int:
             pass
 
         async def f(s):
@@ -92,12 +92,12 @@ class TestTypeAnnotations(TestCase):
 
             @dec
             @overload
-            def f(x):  # type: (int) -> int
+            def f(x: int) -> int:
                 pass
 
             @dec
             @overload
-            def f(x):  # type: (str) -> str
+            def f(x: str) -> str:
                 pass
 
             @dec
@@ -110,11 +110,11 @@ class TestTypeAnnotations(TestCase):
 
         class C:
             @overload
-            def f(self, x):  # type: (int) -> int
+            def f(self, x: int) -> int:
                 pass
 
             @overload
-            def f(self, x):  # type: (str) -> str
+            def f(self, x: str) -> str:
                 pass
 
             def f(self, x): return x
@@ -126,11 +126,11 @@ class TestTypeAnnotations(TestCase):
         import typing as t
 
         @t.overload
-        def f(s):  # type: (None) -> None
+        def f(s: None) -> None:
             pass
 
         @t.overload
-        def f(s):  # type: (int) -> int
+        def f(s: int) -> int:
             pass
 
         def f(s):
@@ -174,7 +174,7 @@ class TestTypeAnnotations(TestCase):
         def f():
             name: str
             age: int
-        ''')
+        ''', m.UnusedAnnotation, m.UnusedAnnotation)
         self.flakes('''
         def f():
             name: str = 'Bob'
@@ -190,7 +190,7 @@ class TestTypeAnnotations(TestCase):
         from typing import Any
         def f():
             a: Any
-        ''')
+        ''', m.UnusedAnnotation)
         self.flakes('''
         foo: not_a_real_type
         ''', m.UndefinedName)
@@ -298,6 +298,11 @@ class TestTypeAnnotations(TestCase):
         a: 'a: "A"'
         ''', m.ForwardAnnotationSyntaxError)
 
+    def test_variable_annotation_references_self_name_undefined(self):
+        self.flakes("""
+        x: int = x
+        """, m.UndefinedName)
+
     def test_TypeAlias_annotations(self):
         self.flakes("""
         from typing_extensions import TypeAlias
@@ -351,11 +356,10 @@ class TestTypeAnnotations(TestCase):
         class Cls:
             y: int
         ''')
-        # TODO: this should print a UnusedVariable message
         self.flakes('''
         def f():
             x: int
-        ''')
+        ''', m.UnusedAnnotation)
         # This should only print one UnusedVariable message
         self.flakes('''
         def f():
@@ -363,13 +367,25 @@ class TestTypeAnnotations(TestCase):
             x = 3
         ''', m.UnusedVariable)
 
+    def test_unused_annotation_in_outer_scope_reassigned_in_local_scope(self):
+        self.flakes('''
+        x: int
+        x.__dict__
+        def f(): x = 1
+        ''', m.UndefinedName, m.UnusedVariable)
+
+    def test_unassigned_annotation_is_undefined(self):
+        self.flakes('''
+        name: str
+        print(name)
+        ''', m.UndefinedName)
+
     def test_annotated_async_def(self):
         self.flakes('''
         class c: pass
         async def func(c: c) -> None: pass
         ''')
 
-    @skipIf(version_info < (3, 7), 'new in Python 3.7')
     def test_postponed_annotations(self):
         self.flakes('''
         from __future__ import annotations
@@ -406,115 +422,6 @@ class TestTypeAnnotations(TestCase):
             __all__: List[str]
         ''')
 
-    def test_typeCommentsMarkImportsAsUsed(self):
-        self.flakes("""
-        from mod import A, B, C, D, E, F, G
-
-
-        def f(
-            a,  # type: A
-        ):
-            # type: (...) -> B
-            for b in a:  # type: C
-                with b as c:  # type: D
-                    d = c.x  # type: E
-                    return d
-
-
-        def g(x):  # type: (F) -> G
-            return x.y
-        """)
-
-    def test_typeCommentsFullSignature(self):
-        self.flakes("""
-        from mod import A, B, C, D
-        def f(a, b):
-            # type: (A, B[C]) -> D
-            return a + b
-        """)
-
-    def test_typeCommentsStarArgs(self):
-        self.flakes("""
-        from mod import A, B, C, D
-        def f(a, *b, **c):
-            # type: (A, *B, **C) -> D
-            return a + b
-        """)
-
-    def test_typeCommentsFullSignatureWithDocstring(self):
-        self.flakes('''
-        from mod import A, B, C, D
-        def f(a, b):
-            # type: (A, B[C]) -> D
-            """do the thing!"""
-            return a + b
-        ''')
-
-    def test_typeCommentsAdditionalComment(self):
-        self.flakes("""
-        from mod import F
-
-        x = 1 # type: F  # noqa
-        """)
-
-    def test_typeCommentsNoWhitespaceAnnotation(self):
-        self.flakes("""
-        from mod import F
-
-        x = 1  #type:F
-        """)
-
-    def test_typeCommentsInvalidDoesNotMarkAsUsed(self):
-        self.flakes("""
-        from mod import F
-
-        # type: F
-        """, m.UnusedImport)
-
-    def test_typeCommentsSyntaxError(self):
-        self.flakes("""
-        def f(x):  # type: (F[) -> None
-            pass
-        """, m.CommentAnnotationSyntaxError)
-
-    def test_typeCommentsSyntaxErrorCorrectLine(self):
-        checker = self.flakes("""\
-        x = 1
-        # type: definitely not a PEP 484 comment
-        """, m.CommentAnnotationSyntaxError)
-        self.assertEqual(checker.messages[0].lineno, 2)
-
-    def test_typeCommentsAssignedToPreviousNode(self):
-        # This test demonstrates an issue in the implementation which
-        # associates the type comment with a node above it, however the type
-        # comment isn't valid according to mypy.  If an improved approach
-        # which can detect these "invalid" type comments is implemented, this
-        # test should be removed / improved to assert that new check.
-        self.flakes("""
-        from mod import F
-        x = 1
-        # type: F
-        """)
-
-    def test_typeIgnore(self):
-        self.flakes("""
-        a = 0  # type: ignore
-        b = 0  # type: ignore[excuse]
-        c = 0  # type: ignore=excuse
-        d = 0  # type: ignore [excuse]
-        e = 0  # type: ignore whatever
-        """)
-
-    def test_typeIgnoreBogus(self):
-        self.flakes("""
-        x = 1  # type: ignored
-        """, m.UndefinedName)
-
-    def test_typeIgnoreBogusUnicode(self):
-        self.flakes("""
-        x = 2  # type: ignore\xc3
-        """, m.UndefinedName)
-
     def test_return_annotation_is_class_scope_variable(self):
         self.flakes("""
         from typing import TypeVar
@@ -533,7 +440,6 @@ class TestTypeAnnotations(TestCase):
                 return Y
         """, m.UndefinedName)
 
-    @skipIf(version_info < (3, 8), 'new in Python 3.8')
     def test_positional_only_argument_annotations(self):
         self.flakes("""
         from x import C
@@ -683,7 +589,6 @@ class TestTypeAnnotations(TestCase):
                 return None
         """)
 
-    @skipIf(version_info < (3, 7), 'new in Python 3.7')
     def test_partial_string_annotations_with_future_annotations(self):
         self.flakes("""
             from __future__ import annotations
@@ -696,6 +601,20 @@ class TestTypeAnnotations(TestCase):
                 return None
         """)
 
+    def test_forward_annotations_for_classes_in_scope(self):
+        # see #749
+        self.flakes("""
+        from typing import Optional
+
+        def f():
+            class C:
+                a: "D"
+                b: Optional["D"]
+                c: "Optional[D]"
+
+            class D: pass
+        """)
+
     def test_idomiatic_typing_guards(self):
         # typing.TYPE_CHECKING: python3.5.3+
         self.flakes("""
@@ -704,7 +623,7 @@ class TestTypeAnnotations(TestCase):
             if TYPE_CHECKING:
                 from t import T
 
-            def f():  # type: () -> T
+            def f() -> T:
                 pass
         """)
         # False: the old, more-compatible approach
@@ -712,7 +631,7 @@ class TestTypeAnnotations(TestCase):
             if False:
                 from t import T
 
-            def f():  # type: () -> T
+            def f() -> T:
                 pass
         """)
         # some choose to assign a constant and do it that way
@@ -722,7 +641,7 @@ class TestTypeAnnotations(TestCase):
             if MYPY:
                 from t import T
 
-            def f():  # type: () -> T
+            def f() -> T:
                 pass
         """)
 
@@ -736,7 +655,7 @@ class TestTypeAnnotations(TestCase):
                 Protocol = object
 
             class C(Protocol):
-                def f():  # type: () -> int
+                def f() -> int:
                     pass
         """)
 
@@ -793,4 +712,88 @@ class TestTypeAnnotations(TestCase):
             def f(*args: *Ts) -> None: ...
 
             def g(x: Shape[*Ts]) -> Shape[*Ts]: ...
+        """)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_statements(self):
+        self.flakes("""
+            type ListOrSet[T] = list[T] | set[T]
+
+            def f(x: ListOrSet[str]) -> None: ...
+
+            type RecursiveType = int | list[RecursiveType]
+
+            type ForwardRef = int | C
+
+            type ForwardRefInBounds[T: C] = T
+
+            class C: pass
+        """)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_functions(self):
+        self.flakes("""
+            def f[T](t: T) -> T: return t
+
+            async def g[T](t: T) -> T: return t
+
+            def with_forward_ref[T: C](t: T) -> T: return t
+
+            def can_access_inside[T](t: T) -> T:
+                print(T)
+                return t
+
+            class C: pass
+        """)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_do_not_escape_function_scopes(self):
+        self.flakes("""
+            from x import g
+
+            @g(T)  # not accessible in decorators
+            def f[T](t: T) -> T: return t
+
+            T  # not accessible afterwards
+        """, m.UndefinedName, m.UndefinedName)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_classes(self):
+        self.flakes("""
+            class C[T](list[T]): pass
+
+            class UsesForward[T: Forward](list[T]): pass
+
+            class Forward: pass
+
+            class WithinBody[T](list[T]):
+                t = T
+        """)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_do_not_escape_class_scopes(self):
+        self.flakes("""
+            from x import g
+
+            @g(T)  # not accessible in decorators
+            class C[T](list[T]): pass
+
+            T  # not accessible afterwards
+        """, m.UndefinedName, m.UndefinedName)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_TypeVarTuple(self):
+        self.flakes("""
+        def f[*T](*args: *T) -> None: ...
+        """)
+
+    @skipIf(version_info < (3, 12), 'new in Python 3.12')
+    def test_type_parameters_ParamSpec(self):
+        self.flakes("""
+        from typing import Callable
+
+        def f[R, **P](f: Callable[P, R]) -> Callable[P, R]:
+            def g(*args: P.args, **kwargs: P.kwargs) -> R:
+                return f(*args, **kwargs)
+            return g
         """)